backend/requirements.txt
flask
flask-sqlalchemy
pandas
python-dotenv

